backend/requirements.txt
fastapi
uvicorn
pydantic
motor
python-dotenv
pydantic-settings
dnspython
passlib[argon2]
python-jose[cryptography]
python-multipart
httpx
google-auth
requests
email-validator
